requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
albumentations>=1.3.0
flask>=2.0.0
matplotlib>=3.5.0
transformers>=4.30.0
pandas>=1.5.0
tqdm>=4.65.0
pillow>=9.0.0
numpy>=1.22.0
